Modify the id of Episode to be equal to the the concatenation of seriesid, season number and episode number.
The episode primary key in db is now the episode id.
The type of episode id is now long, for the concatenation of the three numbers to fit in it.

+++ myseries-unittest/src/mobi/myseries/test/integration/application/image/ImageServiceRepositoryTest.java
@@ -20,7 +20,7 @@ public abstract class ImageServiceRepositoryTest extends InstrumentationTestCase
 
     private static final int TEST_SERIES_ID = -1;
     private static final String TEST_SERIES_NAME = "Test Series";
-    private static final int TEST_EPISODE_ID = -1;
+    private static final long TEST_EPISODE_ID = -1;
     private static final String TEST_EPISODE_NAME = "Test Episode";
 
     private Bitmap testImage;
